requirements.txt
celery==5.3.0
Django==4.2.2
djangorestframework==3.14.0
psycopg2-binary==2.9.6
python-dotenv==1.0.0
redis==4.5.5